Renamed "profile" folder to "user"
